refactor: 함수 이동 및 utils 폴더 삭제
- reissueAccessToken 함수를 src/utils/auth에서 src/api/auth로 이동
- src/utils 폴더 삭제
