+++ src/api/auth.ts
@@ -1,5 +1,7 @@
 import axios from "axios";
 import ResponseData from "../app/models/response-data";
+import useAccessTokenStore from "@/store/accessToken";
+import { redirect } from "next/navigation";
 
 const MOCK_SUCCESS_RESPONSE = {
   data: {
@@ -12,7 +14,7 @@ const MOCK_SUCCESS_RESPONSE = {
   headers: { access: "accessToken" },
 };
 
-export const requestAccessTokenReissue = async () => {
+const requestAccessTokenReissue = async () => {
   // const response = await axios.post<ResponseData>(
   //   "http://localhost:8080/reissue",
   //   {},
@@ -22,4 +24,19 @@ export const requestAccessTokenReissue = async () => {
   const response = MOCK_SUCCESS_RESPONSE;
 
   return response;
+};
+
+export const reissueAccessToken = async () => {
+  const response = await requestAccessTokenReissue();
+
+  if (response.data.code === 200) {
+    const accessToken = response.headers["access"];
+
+    if (accessToken) {
+      const setAccessToken = useAccessTokenStore.getState().setAccessToken;
+      setAccessToken(accessToken);
+    }
+  } else {
+    redirect("/");
+  }
 };
